fix: 2D Monitor fixtures go black above 50% (signed-char RGB truncation)
computeColor() read the RGB channel values with QByteArray::at() straight into
int WITHOUT a uchar cast. at() returns a SIGNED char, so any DMX value > 127
became negative (166 -> -90) and QColor(0,0,-90) clamped the whole colour to
black. Result: every RGB fixture went dark past ~50% intensity, regardless of
model. computeAlpha() and the white/amber folding already cast to uchar; only
the three primary reads were missed.

Root-caused from a file-log capture (qDebug is swallowed by main.cpp's
QtCriticalMsg message handler): the log showed rgb=[0,0,166] => color=#000000,
i.e. a valid blue rendering black. Diagnostic instrumentation removed.

+++ ui/src/monitor/monitorfixtureitem.cpp
@@ -504,9 +504,12 @@ QColor MonitorFixtureItem::computeColor(const FixtureHead *head, const QByteArra
 
     if (head->m_rgb.count() > 0)
     {
-        int r = values.at(head->m_rgb.at(0));
-        int g = values.at(head->m_rgb.at(1));
-        int b = values.at(head->m_rgb.at(2));
+        // uchar cast is essential: QByteArray::at() returns a signed char, so
+        // any DMX value > 127 would become negative and QColor would clamp the
+        // whole colour to black (fixtures went dark past ~50% intensity).
+        int r = uchar(values.at(head->m_rgb.at(0)));
+        int g = uchar(values.at(head->m_rgb.at(1)));
+        int b = uchar(values.at(head->m_rgb.at(2)));
         // Fold in white (adds to all primaries) and amber (warm: R + ~0.5·G)
         // so RGBW/RGBA emitters contribute to the rendered colour.
         foreach (quint32 w, head->m_white)
